refactor(checker): trim alias attribution boundary debt
