+++ crates/tsz-checker/src/query_boundaries/diagnostics.rs
@@ -19,6 +19,15 @@ pub(crate) fn assignment_numeric_display_children(
     tsz_solver::type_queries::assignment_numeric_display_children(db, type_id)
 }
 
+/// `true` when `type_id` is an anonymous object type, or a union / intersection
+/// that contains one (recursing through nested unions / intersections).
+pub(crate) fn union_or_intersection_mentions_object(
+    db: &dyn TypeDatabase,
+    type_id: TypeId,
+) -> bool {
+    tsz_solver::type_queries::union_or_intersection_mentions_object(db, type_id)
+}
+
 pub(crate) fn is_typeof_result_union(db: &dyn TypeDatabase, type_id: TypeId) -> bool {
     const STRING: u8 = 1 << 0;
     const NUMBER: u8 = 1 << 1;
